+++ LeoProjectPayment/src/main/java/com/work/LeoProjectPayment/entity/Transaction.java
@@ -1,15 +1,16 @@
 package com.work.LeoProjectPayment.entity;
 
-import jakarta.persistence.Id;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Column;
-import lombok.Setter;
-import lombok.Getter;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import lombok.ToString;
+
 import java.util.Date;
 
 @Setter
@@ -23,21 +24,21 @@ public class Transaction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Integer id;
 
     @Column(name = "transaction_id")
-    private int transactionId;
+    private Integer transactionId;
 
     @Column(name = "type")
     private String type;
 
     @Column(name = "transaction_amount")
-    private int transactionAmount;
+    private Double transactionAmount;
 
     @Column(name = "date")
     private Date date;
 
-    public Transaction(int transactionId, String type, int transactionAmount, Date date) {
+    public Transaction(int transactionId, String type, Double transactionAmount, Date date) {
         this.transactionId = transactionId;
         this.type = type;
         this.transactionAmount = transactionAmount;
